+++ jib-core/src/main/java/com/google/cloud/tools/jib/image/json/ImageToJsonTranslator.java
@@ -138,6 +138,9 @@ public class ImageToJsonTranslator {
     // Sets the labels.
     template.setContainerLabels(image.getLabels());
 
+    // Sets the working directory.
+    template.setContainerWorkingDir(image.getWorkingDirectory());
+
     // Serializes into JSON.
     return JsonTemplateMapper.toBlob(template);
   }