+++ jib-core/src/main/java/com/google/cloud/tools/jib/image/json/JsonToImageTranslator.java
@@ -162,6 +162,10 @@ public class JsonToImageTranslator {
       }
     }
 
+    if (containerConfigurationTemplate.getContainerLabels() != null) {
+      imageBuilder.setWorkingDirectory(containerConfigurationTemplate.getContainerWorkingDir());
+    }
+
     return imageBuilder.build();
   }